Adds the concept of a HealthServletFactory that allows for the servlet to be configured.
The new servlet behavior defaults to setting a header to indicate that responses should not
be cached. The introduction of a `HealthServletFactory` and `DefaultHealthServletFactory`
supports additional customization of the health check servlet. This includes the ability
to customize the `Cache-Control` header value as well as the ability to
configure the content-type and response body values used for health/unhealthy responses.

The following protected method on the `HealthCheckBundle` is now removed and defining
a `HealthServletFactory` is the new supported mechanism for customising the response
behavior beyond what's supported out of the box.
```
protected HttpServlet createHealthCheckServlet(final AtomicBoolean isHealthy)
```

Also, the `HealthCheckServlet`'s one-argument constructor has been removed in favor of
adding more configurable parameters.

Closes #23